Removed linkedLists.h deppendancy
Got the program to work with out using a linked-list. This is done by restricting the number of cached prime numbers to be no more than 100.  So when the program starts, it makes an array(unsigned long long) of length 100. Once the array reaches the length of 100, the program finds 50 more primes, and sets the last 50 elements of the array to the new prime numbers, in respective order.
The program can now test 1,000,000 every 2.8 seconds.